Relocate compare_feature_classes.py to general_tools directory
The compare_feature_classes.py file has been moved to the general_tools directory within the custom_tools package. All import statements in removing_overlapping_points.py have been updated to recognize this change and maintain proper code functionality.
